requirements.txt
amqp==5.2.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.2
defusedxml==0.8.0rc2
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
django-jazzmin==2.6.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-base64==2.0
drf-yasg==1.21.7
exceptiongroup==1.2.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
inflection==0.5.1
kombu==5.3.5
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyTelegramBotAPI==4.15.4
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-slugify==8.0.4
python-telegram-bot==20.8
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
